Added functions to populate transaction entries and handle item selection.
Added a function to populate the transaction entry fields based on selected transaction in the Treeview (read only display). Added a function to handle updating the transaction entry fields when a transaction is selected from the Treeview (same as functions in other tabs).  Also rearranged a couple blocks of code.